Explicitly cast in npy_2_compat.h for PyDataType_SET_ELSIZE

+++ include/rapidjson/npy_2_compat.h
@@ -145,7 +145,7 @@ PyArray_ImportNumPyAPI()
     static inline void
     PyDataType_SET_ELSIZE(PyArray_Descr *dtype, npy_intp size)
     {
-        dtype->elsize = size;
+        dtype->elsize = static_cast<int>(size);
     }
 
     static inline npy_uint64
